refactor: create regions module
