+++ src/components/forms/club/fields.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 
 import { CreateClubDto, UpdateClubDto } from '@/modules/clubs/types'
 import { CountryDto } from '@/modules/countries/types'
-import { RegionDto } from '@/types/regions'
+import { RegionDto } from '@/modules/regions/types'
 
 import { CountriesCombo } from '../../selects/countries-combo'
 import { RegionsCombo } from '../../selects/regions-combo'
