+++ src/pages/clubs/create.tsx
@@ -4,10 +4,10 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { CreateClubForm } from '@/components/forms/club/create-club'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { useRegionsList } from '@/lib/regions'
 import { withSessionSsr } from '@/lib/session'
 import { useCreateClub } from '@/modules/clubs/hooks'
 import { useCountriesList } from '@/modules/countries/hooks'
+import { useRegionsList } from '@/modules/regions/hooks'
 import { redirectToLogin } from '@/utils/redirect-to-login'
 
 export const getServerSideProps = withSessionSsr(
