+++ src/pages/clubs/edit/[slug].tsx
@@ -5,11 +5,11 @@ import { ErrorContent } from '@/components/error/error-content'
 import { EditClubForm } from '@/components/forms/club/edit-club'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { useRegionsList } from '@/lib/regions'
 import { withSessionSsr } from '@/lib/session'
 import { useUpdateClub } from '@/modules/clubs/hooks'
 import { ClubDto } from '@/modules/clubs/types'
 import { useCountriesList } from '@/modules/countries/hooks'
+import { useRegionsList } from '@/modules/regions/hooks'
 import { getClubBySlug } from '@/services/api/methods/clubs'
 import { ApiError } from '@/types/common'
 import { redirectToLogin } from '@/utils/redirect-to-login'
