+++ src/pages/clubs/index.tsx
@@ -10,11 +10,11 @@ import { ConfirmationModal } from '@/components/modals/confirmation-modal'
 import { PageHeading } from '@/components/page-heading/page-heading'
 import { ClubsTable } from '@/components/tables/clubs'
 import { ClubsTableRow } from '@/components/tables/rows/clubs-row'
-import { useRegionsList } from '@/lib/regions'
 import { withSessionSsr } from '@/lib/session'
 import { useClubs, useDeleteClub } from '@/modules/clubs/hooks'
 import { ClubsFiltersDto, ClubsSortBy } from '@/modules/clubs/types'
 import { useCountriesList } from '@/modules/countries/hooks'
+import { useRegionsList } from '@/modules/regions/hooks'
 import { useLocalStorage } from '@/utils/hooks/use-local-storage'
 import { useTable } from '@/utils/hooks/use-table'
 import { redirectToLogin } from '@/utils/redirect-to-login'
